pyproject.toml
[build-system]
requires = ["setuptools>=69", "setuptools-scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "yapcli"
dynamic = ["version"]
description = "Yet another Plaid CLI."
readme = "README.md"
requires-python = ">=3.12"
authors = [
  { name = "Michael Kane", email = "thisismikekane@gmail.com" }
]
dependencies = [
  "Flask>=3.1.2",
  "itsdangerous>=2.2.0",
  "loguru>=0.7.3",
  "pandas>=2.2.0",
  "platformdirs>=4.3.0",
  "plaid_python>=38.0.0",
  "python-dotenv>=1.2.1",
  "questionary>=2.1.1",
  "typer>=0.12.3",
  "werkzeug>=3.1.4",
]

[project.scripts]
yapcli = "yapcli.cli.main:main"

[project.urls]
Homepage = "https://github.com/thisIsMikeKane/plaid-cli-python"

[tool.setuptools_scm]
version_scheme = "guess-next-dev"
local_scheme = "node-and-date"
fallback_version = "0.0.0"

[dependency-groups]
lint = [
  "black>=24.3",
  "ruff>=0.5.0",
]

typecheck = [
  "mypy>=1.10",
  # Stubs for packages with type hints
  "pandas-stubs>=2.2.0",
  "types-Flask>=1.1.6",
  # Packages with missing type hints
  "loguru>=0.7.3",
  "platformdirs>=4.3.0",
  "python-dotenv>=1.2.1",
  "questionary>=2.1.1",
  "typer>=0.15.3",
]

test = [
  "pytest>=8.3",
  "pytest-cov>=5.0",
  "pytest-sugar>=1.0",
]

tox = [
  "tox>=4.25",
]

build = [
  "build>=1.0",
  "twine>=5.0",
]

dev = [
  "ipykernel>=6.0",
  "pip>=24.2",
  "pre-commit>=3.7",
  { include-group = "lint" },
  { include-group = "typecheck" },
  { include-group = "test" },
  { include-group = "tox" },
  { include-group = "build" },
]

[tool.mypy]
files = ["yapcli"]

[[tool.mypy.overrides]]
module = ["plaid", "plaid.*"]
ignore_missing_imports = true

[tool.setuptools.packages.find]
where = ["."]
include = ["yapcli*"]
exclude = ["tests*"]

[tool.setuptools.package-data]
yapcli = ["frontend/build/**/*"]